fix(ci): 修两条流水线红 —— 导入测试的验证码依赖，和 E2E 连错数据库
两条互不相关，都属于「本地绿、CI 红」：

1. `test_commit_creates_the_users_and_they_can_sign_in` 只设了
   `settings.LOGIN_CAPTCHA_ENABLED = False`，而 `login()` 第一句
   `load_login_config(db)` 会把 sys_config 表里的值 setattr 回 settings。
   那张表在 `fba_test` 里是什么值，取决于上一次 E2E 的 global-setup 跑没跑过 ——
   本地是 `'false'`（跑过），CI 是种子的 `'true'`，于是 CI 上报「验证码无效」，
   跟导入八竿子打不着。`apps/api/AGENTS.md` 早就写着这条，我没照做。

   修法：`test_auth.py` 里那个 `no_captcha` fixture **提到共用 conftest**
   （同 `temp_user` 当初提上来的理由 —— 复制会让 docstring 里那条注意事项分叉）。
   反向验证过：把 `fba_test` 的那条配置改成 `'true'`（= CI 的值）后，
   摘掉 fixture 当场复现 CI 上那句一模一样的错；带上就绿。

2. `.env.e2e.example` 的 `DATABASE_TYPE` 跟着「本地默认库换 PG」一起被改成了
   postgresql，而 CI 的 E2E job 只起 mssql 一个 service → `ConnectionRefusedError`。
   改回 sqlserver（PR #103 本来就写明「test-e2e 仍连 SQL Server」）。

   🔴 顺带查明一件更要紧的事：**CI 里 E2E job 那一整块 `env:` 从来没传进
   api 服务器**。它是 `pnpm e2e`(= turbo) 起的，turbo 2.x 默认 `envMode: strict`，
   而 `turbo.json` 的 e2e 任务没声明 `passThroughEnv` —— 变量在进任务前就被滤掉了
   （实测 `turbo run e2e --dry=json`）。所以 workflow 里那些 DATABASE_* 覆盖
   一直是装饰，真正生效的是 `.env.e2e` 文件，只是以前两边值恰好一样。
   结论记进了 e2e 分册，并纠正了那份文档里「job 级 env 会盖过 dotenv」的旧说法
   —— 那条只对直接跑 pytest 的 job 成立。
