+++ apps/api/backend/.env.e2e.example
@@ -32,17 +32,27 @@ REQUEST_LIMITER_ENABLED=false
 ENVIRONMENT='dev'
 
 # Database —— 指向 fba_test，不是 fba
-DATABASE_TYPE='postgresql'
+#
+# 🔴 **这里刻意还是 sqlserver，跟 `.env.example` 的 postgresql 不一致，别"顺手对齐"。**
+# CI 的 `playwright · E2E` job 只起 mssql 一个 service（`test-e2e` 换方言要单独跑一轮
+# 验证，不跟"改本地默认库"揉在一起 —— PR #103 里写明了这个取舍）。
+#
+# 🔴 而且**这个文件就是权威**，CI job 级的 `env: DATABASE_TYPE/HOST/PORT` 传不进来：
+# E2E 服务器是 `pnpm e2e`(= turbo) 起的，turbo 2.x 默认 `envMode: strict`，
+# 而 `turbo.json` 里 e2e 任务没声明 `passThroughEnv` —— 那些变量在进任务之前就被滤掉了
+# （实测：`turbo run e2e --dry=json` 显示 `envMode: strict` / `passThroughEnv: None`）。
+# 把这里改成 postgresql 之后 CI 当场 `ConnectionRefusedError`（asyncpg 去连 5432，
+# 而那个 job 上只有 1433 的 mssql），而 job 级 env 明明写着 sqlserver —— 看起来完全矛盾。
+DATABASE_TYPE='sqlserver'
 DATABASE_HOST='127.0.0.1'
-DATABASE_PORT=5432
-DATABASE_USER='postgres'
-DATABASE_PASSWORD='postgres'
+DATABASE_PORT=1433
+DATABASE_USER='sa'
+DATABASE_PASSWORD='YourStrong!Passw0rd'
 DATABASE_SOURCES='{}'
 DATABASE_SCHEMA='fba_test'
 DATABASE_PK_MODE='snowflake'
-# 换回 sqlserver 时才需要这两项（postgres/mysql 不看它们）：
-# DATABASE_DRIVER='ODBC Driver 18 for SQL Server'
-# DATABASE_TRUST_SERVER_CERTIFICATE=true
+DATABASE_DRIVER='ODBC Driver 18 for SQL Server'
+DATABASE_TRUST_SERVER_CERTIFICATE=true
 
 # 雪花 ID（沿用开发环境同一套 worker 配置即可，E2E 不会和 dev 实例同时写同一条记录）
 SNOWFLAKE_ENABLED=true
