+++ apps/api/backend/app/admin/tests/api_v1/test_auth.py
@@ -177,29 +177,6 @@ def test_swagger_login_is_not_registered_in_prod(env: str) -> None:
 
 
 @pytest.fixture
-def no_captcha(monkeypatch: pytest.MonkeyPatch) -> Iterator[None]:
-    """让 `/auth/login` 不走验证码
-
-    ⚠️ 光设 `settings.LOGIN_CAPTCHA_ENABLED = False` 不够：`login()` 第一句就是
-    `await load_login_config(db)`，它会把 **sys_config 表里的值** setattr 回 settings
-    （`utils/dynamic_config.py`）。而那张表在 fba_test 里是什么值，取决于上一次
-    E2E 的 global-setup 有没有跑过 —— 依赖它就是依赖一个看不见的外部状态。
-    所以连那次加载一起顶掉。
-    """
-
-    async def _noop(_db: object) -> None:  # ruff: ignore[unused-async] - 顶掉的原函数是 async，签名要对上
-        return None
-
-    monkeypatch.setattr('backend.app.admin.service.auth_service.load_login_config', _noop)
-    original = settings.LOGIN_CAPTCHA_ENABLED
-    settings.LOGIN_CAPTCHA_ENABLED = False
-    try:
-        yield
-    finally:
-        settings.LOGIN_CAPTCHA_ENABLED = original
-
-
-@pytest.fixture
 def session(client: TestClient, no_captcha: None) -> Iterator[tuple[dict[str, str], str]]:
     """走**真实** `/auth/login` 建一个带 refresh cookie 的会话
